% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlm.R
\name{qlm}
\alias{qlm}
\title{Create a SQL query to compute predictions from a GLM.}
\usage{
qlm(model, inverse_link = "{x}")
}
\arguments{
\item{model}{GLM model fit.}

\item{inverse_link}{Inverse link function string, with parameter "{x}".}
}
\value{
SQL query string.
}
\description{
Create a SQL query to compute predictions from a GLM.
}
